Fixed the compilation bugs for file manager. Tommorow do testing on file manager and finish the other functions.

+++ src/file_manager/block.rs
@@ -1,4 +1,4 @@
 pub struct Block_ID{
     pub file_name: String,
-    pub number: u64,
+    pub number: u32,
 }
